// src/apis/cws-api-v2.gen.ts
export namespace CwsApiV2 {
  export const BASE_URL = 'https://chromewebstore.googleapis.com/';

  export type CancelSubmissionRequest = {};

  export type CancelSubmissionResponse = {};

  export type DeployInfo = {
    /**
     * Required. The current deploy percentage for the release channel (nonnegative number between 0 and 100).
     */
    deployPercentage?: number;
  };

  export type DistributionChannel = {
    /**
     * The extension version provided in the manifest of the uploaded package.
     */
    crxVersion?: string;
    /**
     * The current deploy percentage for the release channel (nonnegative number between 0 and 100).
     */
    deployPercentage?: number;
  };

  export type FetchItemStatusResponse = {
    /**
     * Output only. The ID of the item.
     */
    itemId?: string;
    /**
     * Output only. The state of the last async upload for an item. Only set when there has been an async upload for the item in the past 24 hours.
     */
    lastAsyncUploadState?:
      | 'UPLOAD_STATE_UNSPECIFIED'
      | 'SUCCEEDED'
      | 'IN_PROGRESS'
      | 'FAILED'
      | 'NOT_FOUND';
    /**
     * The name of the requested item.
     */
    name?: string;
    /**
     * The public key of the item, which may be generated by the store.
     */
    publicKey?: string;
    /**
     * Output only. Status of the current published revision of the item. Will be unset if the item is not published.
     */
    publishedItemRevisionStatus?: ItemRevisionStatus;
    /**
     * Status of the item revision submitted to be published. Will be unset if the item has not been submitted for publishing since the last successful publish.
     */
    submittedItemRevisionStatus?: ItemRevisionStatus;
    /**
     * If true, the item has been taken down for a policy violation. Check the developer dashboard for details.
     */
    takenDown?: boolean;
    /**
     * If true, the item has been warned for a policy violation and will be taken down if not resolved. Check the developer dashboard for details.
     */
    warned?: boolean;
  };

  export type ItemRevisionStatus = {
    /**
     * Details on the package of the item
     */
    distributionChannels?: Array<DistributionChannel>;
    /**
     * Output only. Current state of the item
     */
    state?:
      | 'ITEM_STATE_UNSPECIFIED'
      | 'PENDING_REVIEW'
      | 'STAGED'
      | 'PUBLISHED'
      | 'PUBLISHED_TO_TESTERS'
      | 'REJECTED'
      | 'CANCELLED';
  };

  export type PublishItemRequest = {
    /**
     * Optional. When set to true the request will fail if there are any warnings during validation and the details will be included in the `error.details`. Otherwise warnings are treated as non-blocking and will be ignored for validation but will be included in the response for inspection. Defaults to `false` if unset.
     */
    blockOnWarnings?: boolean;
    /**
     * Optional. Additional deploy information including the desired initial percentage rollout. Defaults to the current value saved in the developer dashboard if unset.
     */
    deployInfos?: Array<DeployInfo>;
    /**
     * Optional. Use this to control if the item is published immediately on approval or staged for publishing in the future. Defaults to `DEFAULT_PUBLISH` if unset.
     */
    publishType?:
      | 'PUBLISH_TYPE_UNSPECIFIED'
      | 'DEFAULT_PUBLISH'
      | 'STAGED_PUBLISH';
    /**
     * Optional. Whether to attempt to skip item review. The API will validate if the item qualifies and return a validation error if the item requires review. Defaults to `false` if unset.
     */
    skipReview?: boolean;
  };

  export type PublishItemResponse = {
    /**
     * Output only. The ID of the item.
     */
    itemId?: string;
    /**
     * The name of the item that was submitted
     */
    name?: string;
    /**
     * Output only. The current state of the submission.
     */
    state?:
      | 'ITEM_STATE_UNSPECIFIED'
      | 'PENDING_REVIEW'
      | 'STAGED'
      | 'PUBLISHED'
      | 'PUBLISHED_TO_TESTERS'
      | 'REJECTED'
      | 'CANCELLED';
    /**
     * Output only. Non-blocking warnings encountered during the request.
     */
    warningInfo?: WarningsInfo;
  };

  export type SetPublishedDeployPercentageRequest = {
    /**
     * Required. Unscaled percentage value for the publised revision (nonnegative number between 0 and 100). It must be larger than the existing target percentage.
     */
    deployPercentage?: number;
  };

  export type SetPublishedDeployPercentageResponse = {};

  export type UploadItemPackageRequest = {};

  export type UploadItemPackageResponse = {
    /**
     * The extension version provided in the manifest of the uploaded package. This will not be set if the upload is still in progress (`upload_state` is `UPLOAD_IN_PROGRESS`).
     */
    crxVersion?: string;
    /**
     * Output only. The ID of the item the package was uploaded to.
     */
    itemId?: string;
    /**
     * The name of the item the package was uploaded to.
     */
    name?: string;
    /**
     * Output only. The state of the upload. If `upload_state` is `UPLOAD_IN_PROGRESS`, you can poll for updates using the fetchStatus method.
     */
    uploadState?:
      | 'UPLOAD_STATE_UNSPECIFIED'
      | 'SUCCEEDED'
      | 'IN_PROGRESS'
      | 'FAILED'
      | 'NOT_FOUND';
  };
// ...
  export type Endpoints = {
    GET: {
      '/v2/{+name}:fetchStatus': {
        params: {
          /**
           * Required. Name of the item to retrieve the status of in the form `publishers/{publisherId}/items/{itemId}`
           */
          name: string;
        };

        response: { type: 'json'; value: FetchItemStatusResponse };
      };
    };
    POST: {
      '/v2/{+name}:cancelSubmission': {
        params: {
          /**
           * Required. Name of the item to cancel the submission of in the form `publishers/{publisherId}/items/{itemId}`
           */
          name: string;
        };

        body: CancelSubmissionRequest;
        response: { type: 'json'; value: CancelSubmissionResponse };
      };
      '/v2/{+name}:publish': {
        params: {
          /**
           * Required. Name of the item in the form `publishers/{publisherId}/items/{itemId}`
           */
          name: string;
        };

        body: PublishItemRequest;
        response: { type: 'json'; value: PublishItemResponse };
      };
      '/v2/{+name}:setPublishedDeployPercentage': {
        params: {
          /**
           * Required. Name of the item to update the published revision of in the form `publishers/{publisherId}/items/{itemId}`
           */
          name: string;
        };

        body: SetPublishedDeployPercentageRequest;
        response: { type: 'json'; value: SetPublishedDeployPercentageResponse };
      };
      '/upload/v2/{+name}:upload': {
        params: {
          /**
           * Required. Name of the item to upload the new package to in the form `publishers/{publisherId}/items/{itemId}`
           */
          name: string;
        };

        body: Bun.BodyInit;
        response: { type: 'json'; value: UploadItemPackageResponse };
      };
    };
  };
}